The changes in https://github.com/ethereum/go-ethereum/pull/28461 regarding the `sync_test` showed that by adding a resolver, the node iterator sped up considerably. This is really kind of strange, since a memorydb is in essence a map of `string->[]byte`, exactly the same thing as the resolver used instead. 

While investigating, I found that one big reason for the test being slow is that a lot of time is spent on gc. And the upside of using a naive slice instead of memorydb is that memorydb copies on write, and copies again on read. This PR removes one unnecessary copy of the key though. 

```

name           old time/op    new time/op    delta
BatchAllocs-8     412µs ± 6%     382µs ± 2%   -7.18%  (p=0.016 n=5+4)

name           old alloc/op   new alloc/op   delta
BatchAllocs-8     480kB ± 0%     490kB ± 0%   +1.93%  (p=0.008 n=5+5)

name           old allocs/op  new allocs/op  delta
BatchAllocs-8     3.03k ± 0%     2.03k ± 0%  -32.98%  (p=0.008 n=5+5)
```

I also found that the node iterator burns through a lot of memory, which can be remedied by using a local pool. 


```
name        old time/op    new time/op    delta
Iterator-8    6.22ms ± 3%    5.40ms ± 6%  -13.18%  (p=0.008 n=5+5)

name        old alloc/op   new alloc/op   delta
Iterator-8    2.36MB ± 0%    1.67MB ± 0%  -29.23%  (p=0.008 n=5+5)

name        old allocs/op  new allocs/op  delta
Iterator-8     37.0k ± 0%     29.8k ± 0%     ~     (p=0.079 n=4+5)
````

---

There's one remaining source of memory-burn in the node iterator, which comes from the use of `path` separate from node iterator state. This `it.path` is re-allocated on every step, and the node-iterator-state only knows how long their path is, not what the actual path is. I assume this was a decision made to make the node-iterator-state smaller in memory, however, when using a bit of pooling, it would be nicer to move the path into the node-iterator-state, and be able to reuse the `path`-bytes also. Left as a potential future exercise.